cli: move `to_toml_value()` to `src/config.rs`
Google would like to use `to_toml_value()` for writing TOML-formatted
configs from our internal bug report command.
